requirements.txt
python-telegram-bot==13.11
requests==2.27.1
asyncio==3.4.3
PyYAML==6.0